+++ comics/comics/m.py
@@ -4,17 +4,16 @@ from comics.meta.base import MetaBase
 class Meta(MetaBase):
     name = 'M'
     language = 'no'
-    url = 'http://www.madseriksen.no/'
+    url = 'http://www.dagbladet.no/tegneserie/m'
     start_date = '2003-02-10'
     rights = 'Mads Eriksen'
 
 class Crawler(CrawlerBase):
-    has_rerun_releases = True
-    history_capable_date = '2005-01-31'
-    schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
+    history_capable_date = '2011-03-02'
+    schedule = 'Mo,We,Fr'
     time_zone = 1
 
     def crawl(self, pub_date):
-        url = 'http://g2.start.no/tegneserier/striper/m/mstriper/m%s.gif' % (
-            pub_date.strftime('%Y%m%d'),)
+        url = 'http://www.dagbladet.no/tegneserie/markiv/serve.php?%s' % (
+            self.date_to_epoch(pub_date),)
         return CrawlerImage(url)